Ticket #475: Build: Fix Warnings and Compile Errors
git-svn-id: http://svn.osgeo.org/fdo/trunk@4479 d98eaee6-4212-0410-900c-dcc49bf69b99

+++ Providers/SDF/Src/UnitTest/UnitTestUtil.cpp
@@ -393,7 +393,7 @@ void UnitTestUtil::CreateData( bool create, FdoIConnection  *inConn, int featCou
 
 	FdoPtr<FdoPropertyValue> pvNumb = FdoPropertyValue::Create(L"Numb", svNumb);
 
-	FdoDateTime dt(2006,4,21,19,40,10.0001);
+	FdoDateTime dt(2006,4,21,19,40,(float)10.0001);
 	FdoPtr<FdoDateTimeValue> dtv = FdoDateTimeValue::Create(dt);
 	FdoPtr<FdoPropertyValue> pvDateTime = FdoPropertyValue::Create(L"datetime", dtv);
